+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to `ged-api-laravel` will be documented in this file.
 
+## [1.1.0] - 2025-10-09
+
+### Added
+- `CmsBuilder` class - Constrói CMS/PKCS#7 completo com phpseclib3
+- Suporte para montagem correta de SignedData
+- Adicionada dependência `phpseclib/phpseclib ^3.0`
+
+### Fixed
+- CMS agora assina SignedAttributes ao invés de messageDigest
+- Estrutura CMS compatível com Adobe Reader
+
 ## [1.0.0] - 2025-10-08
 
 ### Added